Get movie

+++ src/main/java/pl/krzysiek/model/Movie.java
@@ -3,7 +3,7 @@ package pl.krzysiek.model;
 public class Movie {
     private int id;
     private String release_date;
-    private float vote_average;
+    private double vote_average;
     private String title;
 
     public int getId() {
@@ -22,11 +22,11 @@ public class Movie {
         this.release_date = release_date;
     }
 
-    public float getVote_average() {
+    public double getVote_average() {
         return vote_average;
     }
 
-    public void setVote_average(float vote_average) {
+    public void setVote_average(double vote_average) {
         this.vote_average = vote_average;
     }
